refactor(panel,settings,workspace): clean up redundant code and extract shared UI components
- Extract reusable create_slider_header and bind_debounced_slider in babydra-ui-kit
- Refactor volume and backlight scales in babydra-panel to use shared slider components
- Streamline wifi popover code in babydra-panel: remove redundant parameters, avoid fragile DOM traversing, use update_toggle_state
- Introduce TaskbarPreviewActions struct for workspace preview actions
- Internationalize hardcoded strings in babydra-workspace popover
- Clean up cli argument parsing, collapsible ifs, needless borrows, and static allocations in babydra-settings

+++ crates/babydra-settings/src/widgets/wifi/handler.rs
@@ -207,7 +207,7 @@ pub fn render_network_list(
         icon_badge.set_hexpand(false);
 
         let wifi_icon = babydra_ui_kit::components::create_wifi_net_icon(
-            net.signal as u32,
+            net.signal,
             net.is_connected,
             18,
             Some("#3B82F6"),
